add ntt in macros
